+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/filters/HopByHopHeaderFilter.java
@@ -7,7 +7,6 @@ package de.gematik.test.tiger.mockserver.filters;
 import de.gematik.test.tiger.mockserver.model.Header;
 import de.gematik.test.tiger.mockserver.model.Headers;
 import de.gematik.test.tiger.mockserver.model.HttpRequest;
-import de.gematik.test.tiger.mockserver.model.HttpResponse;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Locale;
@@ -20,24 +19,13 @@ public class HopByHopHeaderFilter {
   private static final List<String> requestHeadersToRemove =
       Arrays.asList(
           "proxy-connection",
-          "connection",
           "keep-alive",
           "transfer-encoding",
           "te",
+          "connection",
           "trailer",
           "proxy-authorization",
           "proxy-authenticate",
-          "upgrade");
-
-  private static final List<String> responseHeadersToRemove =
-      Arrays.asList(
-          "proxy-connection",
-          "connection",
-          "keep-alive",
-          "transfer-encoding",
-          "content-length",
-          "te",
-          "trailer",
           "upgrade");
 
   public HttpRequest onRequest(HttpRequest request) {
@@ -53,24 +41,6 @@ public class HopByHopHeaderFilter {
         clonedRequest.withHeaders(headers);
       }
       return clonedRequest;
-    } else {
-      return null;
-    }
-  }
-
-  public HttpResponse onResponse(HttpResponse response) {
-    if (response != null) {
-      Headers headers = new Headers();
-      for (Header header : response.getHeaderList()) {
-        if (!responseHeadersToRemove.contains(header.getName().toLowerCase(Locale.ENGLISH))) {
-          headers.withEntry(header);
-        }
-      }
-      HttpResponse clonedResponse = response.clone();
-      if (!headers.isEmpty()) {
-        clonedResponse.withHeaders(headers);
-      }
-      return clonedResponse;
     } else {
       return null;
     }