+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/httpclient/HttpClientInitializer.java
@@ -37,7 +37,6 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class HttpClientInitializer extends ChannelInitializer<SocketChannel> {
 
-  private final boolean forwardProxyClient;
   private final Protocol httpProtocol;
   private final HttpClientConnectionErrorHandler httpClientConnectionHandler;
   private final CompletableFuture<Protocol> protocolFuture;
@@ -47,11 +46,9 @@ public class HttpClientInitializer extends ChannelInitializer<SocketChannel> {
 
   HttpClientInitializer(
       Map<ProxyConfiguration.Type, ProxyConfiguration> proxyConfigurations,
-      boolean forwardProxyClient,
       NettySslContextFactory nettySslContextFactory,
       Protocol httpProtocol) {
     this.proxyConfigurations = proxyConfigurations;
-    this.forwardProxyClient = forwardProxyClient;
     this.httpProtocol = httpProtocol;
     this.protocolFuture = new CompletableFuture<>();
     this.httpClientHandler = new HttpClientHandler();
@@ -106,8 +103,7 @@ public class HttpClientInitializer extends ChannelInitializer<SocketChannel> {
       InetSocketAddress remoteAddress = channel.attr(REMOTE_SOCKET).get();
       pipeline.addLast(
           nettySslContextFactory
-              .createClientSslContext(
-                  forwardProxyClient, httpProtocol != null && httpProtocol.equals(Protocol.HTTP_2))
+              .createClientSslContext(httpProtocol != null && httpProtocol.equals(Protocol.HTTP_2))
               .newHandler(channel.alloc(), remoteAddress.getHostName(), remoteAddress.getPort()));
     }