+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/httpclient/NettyHttpClient.java
@@ -59,14 +59,12 @@ public class NettyHttpClient {
   private final Configuration configuration;
   private final EventLoopGroup eventLoopGroup;
   private final Map<ProxyConfiguration.Type, ProxyConfiguration> proxyConfigurations;
-  private final boolean forwardProxyClient;
   private final NettySslContextFactory nettySslContextFactory;
 
   public NettyHttpClient(
       Configuration configuration,
       EventLoopGroup eventLoopGroup,
       List<ProxyConfiguration> proxyConfigurations,
-      boolean forwardProxyClient,
       NettySslContextFactory nettySslContextFactory) {
     this.configuration = configuration;
     this.eventLoopGroup = eventLoopGroup;
@@ -77,7 +75,6 @@ public class NettyHttpClient {
                     Collectors.toMap(
                         ProxyConfiguration::getType, proxyConfiguration -> proxyConfiguration))
             : ImmutableMap.of();
-    this.forwardProxyClient = forwardProxyClient;
     this.nettySslContextFactory = nettySslContextFactory;
   }
 
@@ -123,8 +120,7 @@ public class NettyHttpClient {
           httpRequest.getProtocol() != null ? httpRequest.getProtocol() : Protocol.HTTP_1_1;
 
       final HttpClientInitializer clientInitializer =
-          new HttpClientInitializer(
-              proxyConfigurations, forwardProxyClient, nettySslContextFactory, httpProtocol);
+          new HttpClientInitializer(proxyConfigurations, nettySslContextFactory, httpProtocol);
 
       new Bootstrap()
           .group(eventLoopGroup)
@@ -165,12 +161,7 @@ public class NettyHttpClient {
           (message, throwable) -> {
             if (throwable == null) {
               if (message != null) {
-                if (forwardProxyClient) {
-                  httpResponseFuture.complete(
-                      hopByHopHeaderFilter.onResponse((HttpResponse) message));
-                } else {
-                  httpResponseFuture.complete((HttpResponse) message);
-                }
+                httpResponseFuture.complete((HttpResponse) message);
               } else {
                 httpResponseFuture.complete(response());
               }
@@ -222,8 +213,7 @@ public class NettyHttpClient {
               ERROR_IF_CHANNEL_CLOSED_WITHOUT_RESPONSE,
               !configuration.forwardBinaryRequestsWithoutWaitingForResponse())
           .handler(
-              new HttpClientInitializer(
-                  proxyConfigurations, forwardProxyClient, nettySslContextFactory, null))
+              new HttpClientInitializer(proxyConfigurations, nettySslContextFactory, null))
           .connect(remoteAddress)
           .addListener(
               (ChannelFutureListener)
@@ -298,10 +288,6 @@ public class NettyHttpClient {
       }
     }
     return httpResponse;
-  }
-
-  public HttpResponse sendRequest(HttpRequest httpRequest, long timeout, TimeUnit unit) {
-    return sendRequest(httpRequest, timeout, unit, false);
   }
 
   private boolean isHostNotOnNoProxyHostList(InetSocketAddress remoteAddress) {